prova/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-whois==0.8.0
python-Levenshtein==0.23.0
dnspython==2.4.2
scikit-learn==1.3.2
numpy==1.26.2
joblib==1.3.2
lxml==4.9.3
